pyproject.toml
[project]
name = "amlapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "fpdf2>=2.8.0",
    "httpx>=0.28.0",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[dependency-groups]
dev = [
    "blue>=0.9.1",
]
